+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Changed
 - Include command name in unsupported container command error messages (#1653)
 - Updated arguments name for GCRA command (#1667)
+- Updated INCREX arguments (#1677)
 
 ### Fixed
 - Fixed handling of gap slots in `SlotMap::offsetUnset()` (#1660)